backend/requirements.txt
Flask
flask-cors
opencv-python-headless
numpy
Pillow
scikit-image
pytesseract
